refactor(stellar,backend): extract the consolidated CircuitBreaker into a shared module (#1130)
